Cleaned up remove command
